+++ ml-agents/tests/yamato/yamato_utils.py
@@ -164,7 +164,7 @@ def checkout_csharp_version(csharp_version):
         return
 
     csharp_tag = f"com.unity.ml-agents_{csharp_version}"
-    csharp_dirs = ["com.unity.ml-agents", "com.unity.ml-agents.extensions", "Project"]
+    csharp_dirs = ["com.unity.ml-agents", "Project"]
     for csharp_dir in csharp_dirs:
         subprocess.check_call(f"rm -rf {csharp_dir}", shell=True)
         # Allow the checkout to fail, since the extensions folder isn't availabe in 1.0.0